requirements.txt
Django==4.2.1
djangorestframework==3.14.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.24.0

